ORCH-1145: move venue listing → Hub "Venue" tab (Phase 1, the move only)
Relocate the venue-listing management surface from a row on the founder Brand
profile page into a conditionally-visible "Venue" pill in the Hub, peer to
Events/Experiences/Trips. The management UI is reused as-is (extracted into a
shared VenueListingContent so the Hub tab and the kept route alias render ONE
source). Venue pill shows iff hasPhysicalLocation || placePoolId.

- useHubTabs: HubTabName += venue; deriveHubVisibleTabs gains venue input,
  appends venue LAST (conditional, fails-on-revert); stored-tab guard += venue.
- HubSubNav: HubSubTabId += venue; SUB_TABS/LABELS/ROUTES += venue pill;
  detectActiveSubTab maps /hub/listing → venue.
- New content-only app/(tabs)/hub/listing.tsx: useCurrentBrand() (no param) →
  VenueListingContent chromeMode=tab; forwards ?focus=feedback deep-link.
- _layout: compute venue flags from currentBrand (no 2nd fetch), thread into
  visibility; add /hub/listing → venue branch; PRESERVE nav-lock guard + order.
- brand/[id]/listing.tsx → thin Redirect to /(tabs)/hub/listing (sets active
  brand to route id, forwards ?focus=feedback) — kept for to-do/push/search.
- BrandProfileView + brand/[id]/index: remove the Venue listing row + onListing.
- Tests: new venueTab.contract (T-1..T-10) + listing.orch_1040 re-pointed under
  [TEST-MOD-APPROVED ORCH-1145]; hub-layout-nav-lock preserved unmodified.

No DB/edge/buyer-web/consumer change. experiences.tsx (ORCH-1144) untouched.
