+++ mingla-business/src/components/brand/BrandProfileView.tsx
@@ -191,12 +191,6 @@ export interface BrandProfileViewProps {
    */
   onViewPublic: (brandSlug: string) => void;
   /**
-   * ORCH-1040 — called when user taps the "Venue listing" Operations row.
-   * Receives the brand id; parent routes to /brand/{id}/listing (the venue
-   * listing management page: status, AI scores, changes-remaining, manage).
-   */
-  onListing: (brandId: string) => void;
-  /**
    * Called when user taps the empty-events "Build a new event" CTA.
    * Routes to `/event/create` (the Cycle 3 wedge).
    * NEW in Cycle 7 FX1 — replaces Cycle-2 J-A7 TRANSITIONAL Toast that
@@ -245,7 +239,6 @@ export const BrandProfileView: React.FC<BrandProfileViewProps> = ({
   onAuditLog,
   onBlasts,
   onViewPublic,
-  onListing,
   onCreateEvent,
   onOpenEvent,
   onSeeAllEvents,
@@ -393,27 +386,13 @@ export const BrandProfileView: React.FC<BrandProfileViewProps> = ({
     ? "active"
     : (effectiveStripeStatus ?? brand?.stripeStatus ?? "not_connected");
 
-  // ORCH-1040 — the "Venue listing" row only shows for brands with a physical
-  // location (opt-in flag) OR that already have a linked venue (placePoolId) —
-  // online-only / event brands don't see it until they toggle physical-location on.
-  const showVenueListing =
-    brand?.hasPhysicalLocation === true ||
-    (brand?.placePoolId !== undefined && brand.placePoolId !== null);
-
+  // ORCH-1145 — venue listing moved to the Hub "Venue" tab; do NOT re-add a
+  // brand-page row here (single doorway). The venue-visibility gate now lives
+  // in `deriveHubVisibleTabs` (useHubTabs.ts), conditional on the same
+  // hasPhysicalLocation || placePoolId predicate. The standalone
+  // /brand/{id}/listing route is preserved as a thin redirect to the Hub tab.
   const operationsRows = useMemo<OperationsRow[]>(() => {
     const rows: OperationsRow[] = [];
-    if (showVenueListing) {
-      rows.push({
-        // ORCH-1040 — venue listing management (status, AI scores, changes,
-        // edit/resubmit). First row: the core of the brand's deck presence.
-        icon: "list",
-        label: "Venue listing",
-        sub: "Status, your match scores, and manage",
-        onPress: () => {
-          if (brand !== null) onListing(brand.id);
-        },
-      });
-    }
     rows.push(
       {
         icon: "bank",
@@ -478,8 +457,6 @@ export const BrandProfileView: React.FC<BrandProfileViewProps> = ({
     return rows;
   }, [
     brand,
-    showVenueListing,
-    onListing,
     onTeam,
     onBlasts,
     onPayments,
