+++ mingla-business/src/components/hub/HubSubNav.tsx
@@ -28,7 +28,7 @@ import {
   typography,
 } from "../../constants/designSystem";
 
-export type HubSubTabId = "events" | "experiences" | "trips";
+export type HubSubTabId = "events" | "experiences" | "trips" | "venue";
 export type HubDataDrivenTabId = HubSubTabId | "getstarted";
 
 interface HubSubTab {
@@ -41,6 +41,9 @@ const SUB_TABS: readonly HubSubTab[] = [
   { id: "events", label: "Events", route: "/(tabs)/hub/events" },
   { id: "experiences", label: "Experiences", route: "/(tabs)/hub/experiences" },
   { id: "trips", label: "Trips", route: "/(tabs)/hub/trips" },
+  // ORCH-1145 — conditional Venue pill (rightmost peer; visibility gated in
+  // deriveHubVisibleTabs on hasPhysicalLocation || placePoolId).
+  { id: "venue", label: "Venue", route: "/(tabs)/hub/listing" },
 ] as const;
 
 const LABELS: Record<HubDataDrivenTabId, string> = {
@@ -48,6 +51,7 @@ const LABELS: Record<HubDataDrivenTabId, string> = {
   events: "Events",
   trips: "Trips",
   experiences: "Experiences",
+  venue: "Venue",
 };
 
 const ROUTES: Record<HubDataDrivenTabId, string> = {
@@ -55,6 +59,7 @@ const ROUTES: Record<HubDataDrivenTabId, string> = {
   events: "/(tabs)/hub/events",
   trips: "/(tabs)/hub/trips",
   experiences: "/(tabs)/hub/experiences",
+  venue: "/(tabs)/hub/listing",
 };
 
 export const detectActiveSubTab = (pathname: string): HubDataDrivenTabId => {
@@ -62,6 +67,8 @@ export const detectActiveSubTab = (pathname: string): HubDataDrivenTabId => {
   if (lower.includes("/hub/getstarted")) return "getstarted";
   if (lower.includes("/hub/experiences")) return "experiences";
   if (lower.includes("/hub/trips")) return "trips";
+  // ORCH-1145 — Venue tab route file is `listing.tsx` → URL `/(tabs)/hub/listing`.
+  if (lower.includes("/hub/listing")) return "venue";
   return "events"; // default: Events sub-route is the Hub landing
 };
